requirements.txt
loguru==0.7.2
requests~=2.32.3
nicegui==2.5.0
ruff==0.3.5
fastapi~=0.115.4
uvicorn~=0.32.0
langgraph==0.2.39
loguru==0.7.2
pydantic~=2.8.2
python-dotenv==1.0.1
langchain==0.3.6
langchain-core~=0.3.12
matplotlib~=3.9.2
langchain_community==0.3.4
wikipedia==1.4.0
langchain-fireworks==0.2.5